fix: change to MVC framework
